Move Colormap and ColormapDialog to silx.gui.colors
