+++ doc/source/modules/gui/dialog/colormapdialog.rst
@@ -1,0 +1,12 @@
+
+.. currentmodule:: silx.gui.dialog
+
+:mod:`ColormapDialog`
+=====================
+
+.. automodule:: silx.gui.dialog.ColormapDialog
+
+.. currentmodule:: silx.gui.dialog.ColormapDialog
+
+.. autoclass:: ColormapDialog
+    :members:
