+++ silx/gui/plot/test/testColorBar.py
@@ -26,13 +26,13 @@
 
 __authors__ = ["H. Payno"]
 __license__ = "MIT"
-__date__ = "11/04/2017"
+__date__ = "24/04/2018"
 
 import unittest
 from silx.gui.test.utils import TestCaseQt
 from silx.gui.plot.ColorBar import _ColorScale
 from silx.gui.plot.ColorBar import ColorBarWidget
-from silx.gui.plot.Colormap import Colormap
+from silx.gui.colors import Colormap
 from silx.gui.plot import Plot2D
 from silx.gui import qt
 import numpy
